Extract Other Status Queries To DB
