+++ services/emb_status_service.js
@@ -1,4 +1,7 @@
 "use strict";
+const {updateStatusByCollectionDestinationAndEmbName} = require("../db/embeddings_status");
+const {createStatus} = require("../db/embeddings_status");
+const {getStatusesByCollectionDestinationAndEmbName} = require("../db/embeddings_status");
 const {query, db} = require("@arangodb");
 const {getEmbeddingsFieldName} = require("./emb_collections_service");
 const {embeddingsStatusCollectionName, embeddingsStatus} = require("../model/embeddings_status");
@@ -7,14 +10,7 @@ const {embeddingsStatusCollectionName, embeddingsStatus} = require("../model/emb
  * Get the status of how the embeddings generation is going.
  */
 function getEmbeddingsStatus(collectionName, destinationCollectionName, fieldName, modelMetadata) {
-    const col = db._collection(embeddingsStatusCollectionName);
-    const res = query`
-    FOR d in ${col}
-        FILTER d.collection == ${collectionName}
-        AND d.destination_collection == ${destinationCollectionName}
-        AND d.emb_field_name == ${getEmbeddingsFieldName(fieldName, modelMetadata)}
-        RETURN d
-    `.toArray();
+    const res = getStatusesByCollectionDestinationAndEmbName(collectionName, destinationCollectionName, getEmbeddingsFieldName(fieldName, modelMetadata))
     if (res.length == 0) {
         return embeddingsStatus.DOES_NOT_EXIST;
     }
@@ -25,47 +21,37 @@ function getEmbeddingsStatus(collectionName, destinationCollectionName, fieldNam
  * Get the document ID of an embedding status. Returns `null` if not found.
  */
 function getEmbeddingsStatusDocId(collectionName, destinationCollectionName, fieldName, modelMetadata) {
-    const col = db._collection(embeddingsStatusCollectionName);
-    const res = query`
-    FOR d in ${col}
-        FILTER d.collection == ${collectionName}
-        AND d.destination_collection == ${destinationCollectionName}
-        AND d.emb_field_name == ${getEmbeddingsFieldName(fieldName, modelMetadata)}
-        RETURN d
-    `.toArray();
+    const res = getStatusesByCollectionDestinationAndEmbName(collectionName, destinationCollectionName, getEmbeddingsFieldName(fieldName, modelMetadata))
     if (res.length == 0) {
         return null;
     }
     return res[0]["_key"];
 }
+
 /**
  * Create a new status of how the embeddings generation is going.
  */
 function createEmbeddingsStatus(collectionName, destinationCollectionName, fieldName, modelMetadata) {
-    const col = db._collection(embeddingsStatusCollectionName);
-    query`
-    INSERT {
-        collection: ${collectionName},
-        destination_collection: ${destinationCollectionName},
-        emb_field_name: ${getEmbeddingsFieldName(fieldName, modelMetadata)},
-        status: ${embeddingsStatus.RUNNING}
-    } INTO ${col}
-    `;
+    createStatus(
+        collectionName,
+        destinationCollectionName,
+        getEmbeddingsFieldName(fieldName, modelMetadata),
+        embeddingsStatus.RUNNING
+    );
 }
 
 /**
  * Update the status of how the embeddings generation is going.
  */
 function updateEmbeddingsStatus(newStatus, collectionName, destinationCollectionName, fieldName, modelMetadata) {
-    const col = db._collection(embeddingsStatusCollectionName);
-    query`
-    FOR d in ${col}
-        FILTER d.collection == ${collectionName}
-        AND d.destination_collection == ${destinationCollectionName}
-        AND d.emb_field_name == ${getEmbeddingsFieldName(fieldName, modelMetadata)}
-        UPDATE d._key WITH { status: ${newStatus} } IN ${col}
-    `;
+    updateStatusByCollectionDestinationAndEmbName(
+        collectionName,
+        destinationCollectionName,
+        getEmbeddingsFieldName(fieldName, modelMetadata),
+        newStatus
+    );
 }
+
 exports.getEmbeddingsStatus = getEmbeddingsStatus;
 exports.getEmbeddingsStatusDocId = getEmbeddingsStatusDocId;
 exports.createEmbeddingsStatus = createEmbeddingsStatus;
